+++ lib/presentation/screens/landing_screen.dart
@@ -9,7 +9,7 @@ import 'package:receiptcamp/logic/cubits/landing/landing_cubit.dart';
 import 'package:receiptcamp/logic/cubits/sharing_intent/sharing_intent_cubit.dart';
 import 'package:receiptcamp/presentation/screens/file_explorer.dart';
 import 'package:receiptcamp/presentation/screens/home.dart';
-import 'package:receiptcamp/presentation/screens/recieve_shared_screen.dart';
+import 'package:receiptcamp/presentation/screens/recieve_receipts_screen.dart';
 import 'package:receiptcamp/presentation/ui/landing/app_bar.dart';
 import 'package:receiptcamp/presentation/ui/landing/drawer.dart';
 import 'package:receiptcamp/presentation/ui/landing/nav_bar.dart';
@@ -52,7 +52,7 @@ class _LandingScreenState extends State<LandingScreen>
             case SharingIntentFilesRecieved():
               context.read<SharingIntentCubit>().getFilesAndFolders();
               Navigator.of(context)
-            .push(SlidingReceiptSharingTransitionRoute());
+            .push(SlidingReceiveReceiptTransitionRoute());
             case SharingIntentLoading():
             default:
               return;
